feat: main.js divided into modules
